small changes
added QoL changes for output of findByBooks and findByGenre (better formatting).
made it so only a few fields are required for adding a new book (only isbn, name, and genre) and the rest are optional fields (added curls in comments).
changed path mapping for findbyisbn to camel casing for consistency (findByIsbn).